Run the Prettier code formatter
> prettier --write $(git ls-files)

+++ contracts/algorithms/P256SHA256Algorithm.sol
@@ -5,7 +5,6 @@ import "../BytesUtils.sol";
 import "elliptic-solidity/contracts/curves/EllipticCurve.sol";
 
 contract P256SHA256Algorithm is Algorithm {
-
     using BytesUtils for *;
 
     EllipticCurve public curve;
@@ -21,15 +20,32 @@ contract P256SHA256Algorithm is Algorithm {
     * @param signature The signature to verify.
     * @return True iff the signature is valid.
     */
-    function verify(bytes calldata key, bytes calldata data, bytes calldata signature) external view returns (bool) {
-        return curve.validateSignature(sha256(data), parseSignature(signature), parseKey(key));
+    function verify(
+        bytes calldata key,
+        bytes calldata data,
+        bytes calldata signature
+    ) external view returns (bool) {
+        return
+            curve.validateSignature(
+                sha256(data),
+                parseSignature(signature),
+                parseKey(key)
+            );
     }
 
-    function parseSignature(bytes memory data) internal pure returns (uint256[2] memory) {
-        return [uint256(data.readBytes32(0)), uint256(data.readBytes32(32))];
+    function parseSignature(bytes memory data)
+        internal
+        pure
+        returns (uint[2] memory)
+    {
+        return [uint(data.readBytes32(0)), uint(data.readBytes32(32))];
     }
 
-    function parseKey(bytes memory data) internal pure returns (uint256[2] memory) {
-        return [uint256(data.readBytes32(4)), uint256(data.readBytes32(36))];
+    function parseKey(bytes memory data)
+        internal
+        pure
+        returns (uint[2] memory)
+    {
+        return [uint(data.readBytes32(4)), uint(data.readBytes32(36))];
     }
 }
